+++ Installation/CHANGES.md
@@ -18,6 +18,10 @@
 ### [BGL](https://doc.cgal.org/6.1/Manual/packages.html#PkgBGL)
 -   Added the function `CGAL::Euler::remove_degree_2_vertex()`, which enables users to remove vertices which have exactly two incident edges.
 
+### [3D Convex Hulls](https://doc.cgal.org/6.1/Manual/packages.html#PkgConvexHull3)
+
+- **Breaking change**: For function `CGAL::convex_hull_3()` the traits class has become a named parameter.
+
 ### [2D Arrangements](https://doc.cgal.org/6.1/Manual/packages.html#PkgArrangementOnSurface2)
 
 -   Introduces two traits decorators, namely `Arr_tracing_traits_2` and `Arr_counting_traits_2`, which can be used to extract debugging and informative metadata about the traits in use while a program is being executed.